feat: 콘솔 컨트롤러 생성
