+++ README.md
@@ -11,6 +11,8 @@
   - [x] 기존 출력 기능 코드 가져오기
   - [x] 중간 과정 출력 로직 제거
   - [x] 우승자와 player 별 최종 이동거리 출력
+- [ ] 콘솔 컨트롤러와 웹 컨트롤러 분리
+  - [x] 콘솔 컨트롤러 생성
 
 ## query
 ```sql
